+++ byke-plugin/src/byke/views/layout/Coordinates.java
@@ -7,11 +7,11 @@ public class Coordinates implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	public final float _x;
-	public final float _y;
+	public final int _x;
+	public final int _y;
 
 
-	public Coordinates(float x, float y) {
+	public Coordinates(int x, int y) {
 		_x = x;
 		_y = y;
 	}
@@ -20,7 +20,7 @@ public class Coordinates implements Serializable {
 		return (float)Math.hypot(_x - other._x, _y - other._y);
 	}
 
-	public Coordinates translatedBy(float dx, float dy) {
+	public Coordinates translatedBy(int dx, int dy) {
 		return new Coordinates(_x + dx, _y + dy);
 	}
 }
